requirements.txt
flask
pymongo
prometheus_flask_exporter
prometheus_client
python-dotenv